fix: remove duplicationg command, renaming
